requirements.txt
Django==3.0.7
python-dotenv==0.13.0
django-admin-sortable2==0.7.6
django-extensions==2.2.9
django-tinymce==3.0.2
Pillow==7.1.2
requests==2.24.0
